- Pull script in - Add basic example, update other samples - Add ShowStaticTip and StatusUpdateMessageTimeout params -Add unit tests
