+++ examples/Demo/Server/Pages/_Layout.cshtml
@@ -35,7 +35,5 @@
     <script src="https://cdn.jsdelivr.net/npm/sortablejs@latest/Sortable.min.js"></script>
     <!-- highlight -->
     <script type="text/javascript" src="_content/FluentUI.Demo.Shared/js/highlight-extensions.js"></script>
-    <!-- Touch emulator, included for better browser compatibility for PullToRefresh component -->
-    <!-- <script type="text/javascript" src="_content/FluentUI.Demo.Shared/js/vant-touch-emulator.js"></script> -->
 </body>
 </html>
