+++ examples/Demo/Shared/Pages/AppBar/Examples/AppBarClick.razor
@@ -30,9 +30,6 @@
 </FluentStack>
 
 @code {
-
-    private bool _showSearch = true;
-
     private static Icon HomeIcon(bool active = false) =>
         active ? new Icons.Filled.Size24.Home()
                : new Icons.Regular.Size24.Home();
