+++ examples/Demo/Shared/Pages/PullToRefresh/Examples/PullDownDefault.razor
@@ -1,23 +1,33 @@
-﻿@inject IJSRuntime js
-
-<style>
-    .pull-demo {
+﻿<style>
+    .pull-down-demo {
         height: 51vh;
         max-width: 400px;
         border: calc(var(--stroke-width)* 1px) solid var(--neutral-stroke-rest);
         border-radius: calc(var(--control-corner-radius)* 1px);
-        overflow: auto;
+        overflow-y: scroll;
+        overflow-x: hidden;
     }
 </style>
+<div class="pull-down-demo">
+    <FluentPullToRefresh Direction="@PullDirection.Down" OnRefreshAsync="OnRefreshAsync" ShowStaticTip="false">
+        <PullingTemplate>
+            <FluentIcon Value="@(new Icons.Regular.Size24.ArrowCircleDown())" />
+        </PullingTemplate>
+        <ReleaseTemplate>
+            <FluentIcon Value="@(new Icons.Regular.Size24.ArrowSyncCircle())" />
+        </ReleaseTemplate>
+        <CompletedTemplate>
+            <FluentIcon Value="@(new Icons.Regular.Size24.CheckmarkCircle())" />
+        </CompletedTemplate>
 
-<div class="pull-demo">
-    <FluentPullToRefresh Direction="@PullDirection.Down" OnRefresh="OnRefresh">
-        <div style="user-select: none; display:flex; flex-direction:column; align-items: flex-start; padding: calc(var(--design-unit) * 1px);">
-            @for (int i = 1; i <= count; i++)
-            {
-                <span @key="i">item @i</span>
-            }
-        </div>
+        <ChildContent>
+            <div style="user-select: none; display:flex; flex-direction:column; align-items: flex-start; padding: calc(var(--design-unit) * 1px);">
+                @for (int i = 1; i <= count; i++)
+                {
+                    <span @key="i">item @i</span>
+                }
+            </div>
+        </ChildContent>
     </FluentPullToRefresh>
 </div>
 
@@ -25,22 +35,12 @@
     int refreshcount = 0;
     int count = 20;
 
-    public async Task<bool> OnRefresh()
+    public async Task<bool> OnRefreshAsync()
     {
         refreshcount++;
         DemoLogger.WriteLine($"Pull down refresh count {refreshcount}");
         await Task.Delay(1000);
         count += 20;
         return true;
-    }
-
-    protected override async Task OnAfterRenderAsync(bool firstRender)
-    {
-        if (firstRender)
-        {
-            // Load touch emulator. This is included to enable touch events on desktop and to add compatibility
-            // with all browsers. It is not needed for the actual component.
-            //await js.InvokeVoidAsync("loadTouchEmulator");
-        }
     }
 }
