+++ examples/Demo/Shared/Pages/PullToRefresh/Examples/PullUpDefault.razor
@@ -1,6 +1,6 @@
 ﻿@inject IJSRuntime js
 <style>
-    .pull-demo {
+    .pull-up-demo {
         height: 51.2vh;
         max-width: 400px;
         border: calc(var(--stroke-width)* 1px) solid var(--neutral-stroke-rest);
@@ -16,19 +16,19 @@
     }
 </style>
 
-<div class="pull-demo">
-    <FluentPullToRefresh Direction="@PullDirection.Up" OnRefresh="OnRefresh" PullTipHeight="32">
+<div class="pull-up-demo">
+    <FluentPullToRefresh Direction="@PullDirection.Up" OnRefreshAsync="OnRefreshAsync" TipHeight="50">
         <PullingTemplate>Pull up to refresh</PullingTemplate>
         <LoadingTemplate>
             <FluentProgress Width="150px;"/>
         </LoadingTemplate>
         <ChildContent>
-        <div class="pull-content">
-            @for (int i = 1; i <= count; i++)
-            {
-                <span @key="i">item @i</span>
-            }
-        </div>
+            <div class="pull-content">
+                @for (int i = 1; i <= count; i++)
+                {
+                    <span @key="i">item @i</span>
+                }
+            </div>
         </ChildContent>
     </FluentPullToRefresh>
 </div>
@@ -37,7 +37,7 @@
     int refreshcount = 0;
     int count = 20;
 
-    public async Task<bool> OnRefresh()
+    public async Task<bool> OnRefreshAsync()
     {
         refreshcount++;
         if (count < 100)
@@ -50,15 +50,5 @@
         }
 
         return false;
-    }
-
-    protected override async Task OnAfterRenderAsync(bool firstRender)
-    {
-        if (firstRender)
-        {
-            // Load touch emulator. This is included to enable touch events on desktop and to add compatibility
-            // with all browsers. It is not needed for the actual component.
-           // await js.InvokeVoidAsync("loadTouchEmulator");
-        }
     }
 }
