+++ examples/Demo/Shared/Pages/PullToRefresh/PullToRefreshPage.razor
@@ -1,6 +1,19 @@
 ﻿@page "/PullToRefresh"
 @using FluentUI.Demo.Shared.Pages.PullToRefresh.Examples
 
+<p>
+    The <b>Pull-to-refresh</b> is a touchscreen gesture. It involves touching the screen of a computing device with a finger
+    or pressing a button on a pointing device, dragging the screen downward, and then releasing it.
+    This action signals the application to refresh the contents of the component.
+    Its purpose is to make refreshing immediately accessible and to save valuable screen space that would otherwise be occupied by a refresh button.
+</p>
+
+<p>
+    These features are mainly used on mobile devices. To maintain compatibility with the majority of <b>desktop browsers</b>, an emulator script is included 
+    and loaded automatically by the component.
+</p>
+<DemoSection Title="Default" Component="typeof(PullDownBasic)"></DemoSection>
+
 <DemoSection Title="Pull down" Component="typeof(PullDownDefault)"></DemoSection>
 
 <DemoSection Title="Pull up" Component="typeof(PullUpDefault)">
