+++ src/Core.Assets/src/index.ts
@@ -2,7 +2,6 @@ export * from '@fluentui/web-components/dist/web-components'
 import { SplitPanels } from './SplitPanels'
 import { DesignTheme } from './DesignTheme'
 import { FluentPageScript, onEnhancedLoad } from './FluentPageScript'
-import { TouchEmulator } from './TouchEmulator' 
 
 interface Blazor {
   registerCustomEventType: (
@@ -306,8 +305,6 @@ export function afterStarted(blazor: Blazor, mode: string) {
 export function beforeStart(options: any) {
   customElements.define("fluent-design-theme", DesignTheme);
   customElements.define("split-panels", SplitPanels);
-
-  new TouchEmulator();
 
   beforeStartCalled = true;
 }
