+++ src/Core/Components/PullToRefresh/FluentPullToRefresh.razor.css
@@ -3,28 +3,24 @@
     flex-direction: column;
     flex-grow: 1;
 }
-    .fluent-pull-container > .fluent-pull-wrapper {
-        position: relative;
+
+    .fluent-pull-container div[part="tip"] {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        background-color: var(--neutral-layer-2);
+        color: var(--neutral-base-color);
+        padding: 0.5rem;
+        height: var(--fluent-pull-refresh-head-height, 50px);
+        width: 100%;
     }
 
-.fluent-pull-tip {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: var(--neutral-layer-2);
-    color: var(--neutral-base-color);
-    padding: 0.5rem;
-    height: var(--fluent-pull-refresh-head-height, 50px);
-    width: 100%;
-}
+        .fluent-pull-container div[part="tip"][direction="down"] {
+            position: relative;
+            top: 0px;
+        }
 
-.fluent-pull-down-tip {
-    position: relative;
-    top: 0px;
-
-}
-
-.fluent-pull-up-tip {
-    position: relative;
-    bottom: 0px;
-}
+        .fluent-pull-container div[part="tip"][direction="up"] {
+            position: relative;
+            bottom: 0px;
+        }
