+++ tests/Core/PullToRefresh/FluentPullToRefreshTests.FluentPullToRefresh_DirectionUp.verified.razor.html
@@ -1,0 +1,4 @@
+﻿<div class="fluent-pull-container" blazor:ontouchstart="1" blazor:ontouchmove="2" blazor:ontouchend="3" b-rq0hhda1ik="">
+  <div style="position: relative; user-select: none;" b-rq0hhda1ik="">Content that can be refreshed<div part="tip" direction="up" style="--fluent-pull-refresh-head-height: 32px;" b-rq0hhda1ik="">Pull down to refresh</div>
+  </div>
+</div>
