+++ tests/Core/PullToRefresh/FluentPullToRefreshTests.FluentPullToRefresh_DoNotShowStaticTip.verified.razor.html
@@ -1,0 +1,3 @@
+﻿<div class="fluent-pull-container" blazor:ontouchstart="1" blazor:ontouchmove="2" blazor:ontouchend="3" b-rq0hhda1ik="">
+  <div style="position: relative; user-select: none;" b-rq0hhda1ik="">Content that can be refreshed</div>
+</div>
