+++ tests/Core/PullToRefresh/FluentPullToRefreshTests.FluentPullToRefresh_LoadingTemplate.verified.razor.html
@@ -1,0 +1,4 @@
+﻿<div class="fluent-pull-container" blazor:ontouchstart="1" blazor:ontouchmove="2" blazor:ontouchend="3" b-rq0hhda1ik="">
+  <div style="position: relative; user-select: none;" b-rq0hhda1ik="">
+    <div part="tip" direction="down" style="--fluent-pull-refresh-head-height: 32px;" b-rq0hhda1ik="">Pull to refresh...</div>Content that can be refreshed</div>
+</div>
